Split timeseries samples (#1578)
* Split timeseries samples

* Resolve conversations
